requirements.txt
Django==1.10
django-carton==1.2.1
Pillow==3.3.0
pytz==2016.6.1
sorl-thumbnail==12.3
